requirements.txt
pynput==1.7.6
pyautogui==0.9.54
pyperclip==1.8.2
openai==1.13.3
python-dotenv==1.0.1